Fix some issues with decors and foods
- Update FoodAPIView, DecorAPIView and the photos views

I was able to put a photo that does not belong to a food as main photo and same thing for decors
and there was other errors

+++ EventEase_server/photos/views.py
@@ -8,7 +8,7 @@ from rest_framework.generics import DestroyAPIView
 from rest_framework.response import Response
 from rest_framework.parsers import MultiPartParser
 from rest_framework.permissions import IsAuthenticated
-from services.models import Service,Decor
+from services.models import Service,Decor,FoodType,FoodServiceFood
 from accounts.permissions import IsServiceOwner, IsServiceOwnerOrAdmin
 from . models import ServicePhotos,ServiceProfilePhoto,FoodPhotos,Food,MainFoodPhoto,DecorPhotos,MainDecorPhoto
 from . serializers import (ServicePhotosSerializers,ServiceProfilePhotoSerialzer,FoodPhotosSerializers,
@@ -111,25 +111,42 @@ class FoodPhotosAPIView(APIView):
     parser_classes = [MultiPartParser]
 
     def get(self, request, service_pk, type_pk, food_pk, photo_pk=None):
+        service = get_object_or_404(Service,id = service_pk)
+        food_type = get_object_or_404(FoodType, id = type_pk)
         food = get_object_or_404(Food, id=food_pk)
+        food_service_food = FoodServiceFood.objects.filter(food=food, foodService = service, food__food_type = food_type).first()
+        if food_service_food is None:
+            return Response({'message':'no such food in this service'},status= 400)
         if photo_pk:
-            photo = get_object_or_404(FoodPhotos, id=photo_pk, food=food)
+            photo = get_object_or_404(FoodPhotos, id=photo_pk, food=food_service_food.food)
             serializer = FoodPhotosSerializers(photo)
             return Response(serializer.data)
-        photos = FoodPhotos.objects.filter(food=food)
+        photos = FoodPhotos.objects.filter(food=food_service_food.food)
         serializer = FoodPhotosSerializers(photos, many=True)
         return Response(serializer.data)
 
     def post(self, request, service_pk, type_pk, food_pk):
+        service = get_object_or_404(Service,id = service_pk)
+        food_type = get_object_or_404(FoodType, id = type_pk)
         food = get_object_or_404(Food, id=food_pk)
+        food_service_food = FoodServiceFood.objects.filter(food=food, foodService = service, food__food_type = food_type).first()
+        print(food)
+        if food_service_food is None:
+            return Response({'message':'no such food in this service'},status= 400)
         serializer = FoodPhotosSerializers(data=request.data)
         if serializer.is_valid():
-            serializer.save(food=food)
+            serializer.save(food=food_service_food.food)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, service_pk, type_pk, food_pk, photo_pk):
+        service = get_object_or_404(Service,id = service_pk)
+        food_type = get_object_or_404(FoodType, id = type_pk)
+        food = get_object_or_404(Food, id=food_pk)
+        food_service_food = FoodServiceFood.objects.filter(food=food, foodService = service, food__food_type = food_type).first()
+        if food_service_food is None:
+            return Response({'message':'no such food in this service'},status= 400)
         photo = get_object_or_404(FoodPhotos, id=photo_pk)
         image_path = str(photo.image)
         try:
@@ -152,22 +169,43 @@ class FoodPhotosAPIView(APIView):
 class MainFoodPhotoAPIView(APIView):
 
     def get(self, request, service_pk, type_pk, food_pk):
+        service = get_object_or_404(Service,id = service_pk)
+        food_type = get_object_or_404(FoodType, id = type_pk)
+        food = get_object_or_404(Food, id=food_pk)
+        food_service_food = FoodServiceFood.objects.filter(food=food, foodService = service, food__food_type = food_type).first()
+        if food_service_food is None:
+            return Response({'message':'no such food in this service'},status= 400)
         main_photo = get_object_or_404(MainFoodPhoto, food_id=food_pk)
         serializer = MainFoodPhotoSerializer(main_photo)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request, service_pk, type_pk, food_pk):
+        service = get_object_or_404(Service,id = service_pk)
+        food_type = get_object_or_404(FoodType, id = type_pk)
         food = get_object_or_404(Food, id=food_pk)
+        food_service_food = FoodServiceFood.objects.filter(food=food, foodService = service, food__food_type = food_type).first()
+        if food_service_food is None:
+            return Response({'message':'no such food in this service'},status= 400)
         serializer = MainFoodPhotoSerializer(data=request.data)
         if serializer.is_valid():
+            print(serializer.validated_data.get('foodPhoto'))
+            food_photo = serializer.validated_data.get('foodPhoto')
+            if food_service_food.food != food_photo.food: # if the photo that we're going to set main does not belong to the food
+                return Response({'message':'the photo you chosed deos not belong to the corresponding food'},status=400)
             main_photo, created = MainFoodPhoto.objects.update_or_create(
-                food=food,
+                food=food_service_food.food,
                 defaults={'foodPhoto': serializer.validated_data['foodPhoto']}
             )
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, service_pk, type_pk, food_pk):
+        service = get_object_or_404(Service,id = service_pk)
+        food_type = get_object_or_404(FoodType, id = type_pk)
+        food = get_object_or_404(Food, id=food_pk)
+        food_service_food = FoodServiceFood.objects.filter(food=food, foodService = service, food__food_type = food_type).first()
+        if food_service_food is None:
+            return Response({'message':'no such food in this service'},status= 400)
         main_photo = get_object_or_404(MainFoodPhoto, food_id=food_pk)
         main_photo.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
@@ -185,7 +223,8 @@ class DecorPhotosAPIView(APIView):
     parser_classes = [MultiPartParser]
 
     def get(self, request, service_pk, decor_pk, photo_pk=None):
-        decor = get_object_or_404(Decor, id=decor_pk)
+        service = get_object_or_404(Service,id = service_pk)
+        decor = get_object_or_404(Decor, id=decor_pk, decor_service = service) 
         if photo_pk:
             photo = get_object_or_404(DecorPhotos, id=photo_pk, decor=decor)
             serializer = DecorPhotosSerializer(photo)
@@ -195,7 +234,8 @@ class DecorPhotosAPIView(APIView):
         return Response(serializer.data)
 
     def post(self, request, service_pk, decor_pk):
-        decor = get_object_or_404(Decor, id=decor_pk)
+        service = get_object_or_404(Service,id = service_pk)
+        decor = get_object_or_404(Decor, id=decor_pk, decor_service = service)
         serializer = DecorPhotosSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save(decor=decor)
@@ -204,6 +244,8 @@ class DecorPhotosAPIView(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, service_pk, decor_pk, photo_pk):
+        service = get_object_or_404(Service,id = service_pk)
+        decor = get_object_or_404(Decor, id=decor_pk, decor_service = service)
         photo = get_object_or_404(DecorPhotos, id=photo_pk)
         image_path = str(photo.image)
         try:
@@ -225,14 +267,20 @@ class DecorPhotosAPIView(APIView):
 class MainDecorPhotoAPIView(APIView):
 
     def get(self, request, service_pk, decor_pk):
-        main_photo = get_object_or_404(MainDecorPhoto, decor_id=decor_pk)
+        service = get_object_or_404(Service,id = service_pk)
+        decor = get_object_or_404(Decor, id=decor_pk, decor_service = service)
+        main_photo = get_object_or_404(MainDecorPhoto, decor=decor)
         serializer = MainDecorPhotoSerializer(main_photo)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request, service_pk, decor_pk):
-        decor = get_object_or_404(Decor, id=decor_pk)
+        service = get_object_or_404(Service,id = service_pk)
+        decor = get_object_or_404(Decor, id=decor_pk, decor_service = service)
         serializer = MainDecorPhotoSerializer(data=request.data)
         if serializer.is_valid():
+            decorPhoto = serializer.validated_data.get('decorPhoto')
+            if decor != decorPhoto.decor: # if the photo that we're going to set main does not belong to the decor
+                return Response({'message':'the photo you chosed deos not belong to the corresponding decor'},status=400)
             main_photo, created = MainDecorPhoto.objects.update_or_create(
                 decor=decor,
                 defaults={'decorPhoto': serializer.validated_data['decorPhoto']}
@@ -241,7 +289,9 @@ class MainDecorPhotoAPIView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, service_pk, decor_pk):
-        main_photo = get_object_or_404(MainDecorPhoto, decor_id=decor_pk)
+        service = get_object_or_404(Service,id = service_pk)
+        decor = get_object_or_404(Decor, id=decor_pk, decor_service = service)
+        main_photo = get_object_or_404(MainDecorPhoto, decor=decor)
         main_photo.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
